Use persistent's SqlPersistM
This is for #19. We'll see what the test suite says. I'm not really
sure why but we were unrolling a transformer stack
ResourceT (NoLoggingT IO) but the one that persistent uses is
NoLoggingT (ResourceT IO). The fact that they alias to SqlPersistM as
an exported type alias is a strong signal that we should be using it
and I can't think of a reason why we would need those two layers of
the stack swapped.